Fix cast warning. Now count is 131

+++ src/kbd.cc
@@ -1416,7 +1416,7 @@ static int
 get_kbd_layout_name (HKL hkl, char *buf, int size)
 {
   char k[256];
-  sprintf (k, "SYSTEM\\CurrentControlSet\\Control\\Keyboard Layouts\\%16llx", hkl);
+  sprintf (k, "SYSTEM\\CurrentControlSet\\Control\\Keyboard Layouts\\%16llx", (unsigned long long) hkl);
   ReadRegistry r (HKEY_LOCAL_MACHINE, k);
   return ((!r.fail () && r.get ("Layout Text", buf, size) > 0)
           || app.kbdq.gime.ImmGetDescription (hkl, buf, size) > 0);
